@@ -17,25 +17,25 @@
 		"@iconify/svelte": "^3.1.6",
 		"@playwright/test": "^1.28.1",
 		"@sveltejs/adapter-auto": "^3.1.1",
-		"@sveltejs/kit": "^2.5.2",
+		"@sveltejs/kit": "^2.5.27",
 		"@typescript-eslint/eslint-plugin": "^5.45.0",
 		"@typescript-eslint/parser": "^5.45.0",
 		"autoprefixer": "^10.4.17",
 		"babel-plugin-prismjs": "^2.1.0",
 		"eslint": "^8.28.0",
 		"eslint-config-prettier": "^8.5.0",
-		"eslint-plugin-svelte": "^2.30.0",
+		"eslint-plugin-svelte": "^2.45.1",
 		"happy-dom": "^9.20.3",
 		"postcss": "^8.4.35",
-		"prettier": "^2.8.0",
-		"prettier-plugin-svelte": "^2.10.1",
-		"svelte": "^4.0.0",
-		"svelte-check": "^3.4.3",
+		"prettier": "^3.1.0",
+		"prettier-plugin-svelte": "^3.2.6",
+		"svelte": "^5.0.0",
+		"svelte-check": "^4.0.0",
 		"svelte-spa-router": "^3.3.0",
 		"tailwindcss": "^3.4.1",
 		"tslib": "^2.4.1",
-		"typescript": "^5.0.0",
-		"vite": "^5.1.4",
+		"typescript": "^5.5.0",
+		"vite": "^5.4.4",
 		"vitest": "^0.25.3"
 	},
 	"type": "module",
