@@ -10,7 +10,7 @@
 </svelte:head>
 
 <Header />
-<div class="mb-16" />
+<div class="mb-16"></div>
 <main class="container mx-auto px-4 py-12 max-w-2xl">
 	<div class="error-box m-6" role="alert">
 		<p>{$page.status}: {$page.error?.message}</p>
